VS2008环境下消除告警和编译错误

+++ DMC/DmcManager.h
@@ -73,7 +73,7 @@ public:
 	unsigned long init();
 	void close();
 	//记录CSP规划点位置到日志文件中
-	void logCspPoints(const Item *pItems, int rows, int cols) const;
+	void logCspPoints(const Item *pItems, int rows, size_t cols) const;
 
 	//运动
 	unsigned long clr_alarm(short axis);
